move zeroex to separate file
